py/apicomposer/pyproject.toml
[project]
name = "apicomposer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.58.1",
    "python-dotenv>=1.0.1",
    "grpcio>=1.59.0",
    "grpcio-tools>=1.59.0"
]

[tool.poetry]
name = "apicomposer"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
